* variable.c (rb_autoload_load): suppress a warning.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@23374 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ variable.c
@@ -1440,7 +1440,7 @@ rb_autoload_load(VALUE klass, ID id)
 
     if (!load) return Qfalse;
     file = load->nd_lit;
-    return rb_require_safe(file, load->nd_nth);
+    return rb_require_safe(file, (int)load->nd_nth);
 }
 
 VALUE
